1182-Separate solutions
